+++ modules/apps/segments/segments-experiment-web/src/main/resources/META-INF/resources/js/util/navigation.es.js
@@ -1,0 +1,36 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+/**
+ * Generates standard navigation between Experiences
+ *
+ * @export
+ * @param {string} experienceId
+ * @param {string} [baseUrl=window.location.href]
+ */
+export function navigateToExperience(
+	experienceId,
+	baseUrl = window.location.href
+) {
+	const currentUrl = new URL(baseUrl);
+	const urlQueryString = currentUrl.search;
+	const urlSearchParams = new URLSearchParams(urlQueryString);
+
+	urlSearchParams.set('segmentsExperienceId', experienceId);
+	currentUrl.search = urlSearchParams.toString();
+
+	const newUrl = currentUrl.toString();
+
+	Liferay.Util.navigate(newUrl);
+}
